Исправлено:     Приложение ломается при перезагрузке страницы с открытым попапом https://disk.yandex.ru/i/U-celuGErh7OBA     Код хука нужно поместить в отдельный файл useForm.js в папке hooks и импортируют функцию туда, где нужно контролировать инпуты     dispatch(getIngredients()) достаточно 1 раз сделать в App, чтобы все страницы получили ингредиенты при загрузке сайта. Не нужно дублировать запрос в других компонентах.
