+++ src/pages/ForgotPassword/ForgotPassword.jsx
@@ -4,7 +4,8 @@ import { restorePassword } from '../../services/actions/auth/auth';
 import { Input } from '@ya.praktikum/react-developer-burger-ui-components';
 import { Link, Navigate, useLocation } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import {Form, useForm} from '../../components/form/Form';
+import {Form} from '../../components/form/Form';
+import { useForm } from '../../hooks/useForm';
 
 export function ForgotPassword() {
   const {values, handleChange } = useForm({ email: '' });
