+++ src/pages/HomePage/homePage.jsx
@@ -4,19 +4,12 @@ import BurgerIngredients from "../../components/burger-ingredients/burger-ingred
 import BurgerConstructor from "../../components/burger-constructor/burger-constructor";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-import { getIngredients } from "../../services/actions/burger-ingredients/burger-ingredients";
-import { useSelector, useDispatch } from "react-redux";
-import React, { useEffect } from "react";
+import { useSelector } from "react-redux";
 
 export function HomePage() {
   const { ingredientsChecker, ingredientsRequest, ingredients } = useSelector(
     (state) => state.burgerIngredientsReducer
   );
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(getIngredients());
-  }, [dispatch]);
 
   if (ingredientsRequest) {
     return <h2>Загрузка...</h2>;
